Order 생성 기능 (테스트 필요)

+++ src/main/java/net/softsociety/bacs/menu/dto/UpdateMenuOptionDTO.java
@@ -14,7 +14,7 @@ public record UpdateMenuOptionDTO(
         @NotBlank
         String optionValue,
         @NotBlank
-        long optionPrice,
+        int optionPrice,
         @NotBlank
         Menu menu_id
 ) {}
